applications/account/views.py
from django.shortcuts import render
from rest_framework.generics import CreateAPIView
from django.contrib.auth import get_user_model
from applications.account.serializers import RegisterSerializer
# Create your views here.


User = get_user_model()


class UserRegisterAPIView(CreateAPIView):
  queryset = User.objects.all()
  serializer_class = RegisterSerializer